+++ perl/FSM/IR/LoweredRTLIR.pm
@@ -17,6 +17,9 @@ sub new ($class, %args) {
         target_language => $args{target_language} // 'systemverilog',
         output_drive_families => _clone($args{output_drive_families} || []),
         standalone_dt_multi_drive_targets => _clone($args{standalone_dt_multi_drive_targets} || []),
+        internal_net_names => _clone($args{internal_net_names}),
+        instance_names => _clone($args{instance_names}),
+        auxiliary_assignment_count => $args{auxiliary_assignment_count},
     }, $class;
 }
 
@@ -25,12 +28,17 @@ sub source_root_kind ($self) { return $self->{source_root_kind} }
 sub target_language ($self) { return $self->{target_language} }
 sub output_drive_families ($self) { return $self->{output_drive_families} }
 sub standalone_dt_multi_drive_targets ($self) { return $self->{standalone_dt_multi_drive_targets} }
+sub internal_net_names ($self) { return $self->{internal_net_names} }
+sub instance_names ($self) { return $self->{instance_names} }
+sub auxiliary_assignment_count ($self) { return $self->{auxiliary_assignment_count} }
 
 sub as_hashref ($self) {
     my $output_drive_families = _clone($self->output_drive_families || []);
     my $standalone_dt_multi_drive_targets = _clone($self->standalone_dt_multi_drive_targets || []);
+    my $internal_net_names = _clone($self->internal_net_names);
+    my $instance_names = _clone($self->instance_names);
 
-    return {
+    my $result = {
         module_name => $self->module_name,
         source_root_kind => $self->source_root_kind,
         target_language => $self->target_language,
@@ -39,6 +47,21 @@ sub as_hashref ($self) {
         standalone_dt_multi_drive_target_count => scalar(@$standalone_dt_multi_drive_targets),
         standalone_dt_multi_drive_targets => $standalone_dt_multi_drive_targets,
     };
+
+    if (defined $internal_net_names) {
+        $result->{internal_net_count} = scalar(@{$internal_net_names || []});
+        $result->{internal_net_names} = $internal_net_names;
+    }
+
+    if (defined $instance_names) {
+        $result->{instance_count} = scalar(@{$instance_names || []});
+        $result->{instance_names} = $instance_names;
+    }
+
+    $result->{auxiliary_assignment_count} = $self->auxiliary_assignment_count
+        if defined $self->auxiliary_assignment_count;
+
+    return $result;
 }
 
 sub _clone ($value) {